pyproject.toml
[project]
name = "pixelplay"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "streamlit>=1.50.0",
    "pillow>=10.0.0",
    "numpy>=1.26.0",
    "opencv-python-headless>=4.8.0",
]

[dependency-groups]
dev = ["tqdm>=4.67.1", "watchdog>=6.0.0"]
